accessible/tests/mochitest/treeupdate/test_table.html

<!DOCTYPE html>
<html>
<head>
  <title>Table update tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">

    async function doTest() {
      let reordered = waitForEvent(EVENT_REORDER, "table");
      // append a caption, it should appear as a first element in the
      // accessible tree.
      const caption = document.createElement("caption");
      caption.textContent = "table caption";
      getNode("table").appendChild(caption);
      await reordered;

      const tree =
        { TABLE: [
          { CAPTION: [
            { TEXT_LEAF: [] },
          ] },
          { ROW: [
            { CELL: [{ TEXT_LEAF: [] }] },
            { CELL: [{ TEXT_LEAF: [] }] },
          ] },
        ] };
      testAccessibleTree("table", tree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <table id="table">
    <tr>
      <td>cell1</td>
      <td>cell2</td>
    </tr>
  </table>
</body>
</html>
